+++ R/pq_return.R
@@ -26,17 +26,15 @@ dat_add_byfreq = function(dat, freq) {
     return(dat2)
 }
 
-pq1_return = function(dt, x, freq='monthly', num=1, date_type = 'eop', method = 'arithmetic', leading=TRUE, cumreturns = FALSE, rcol_name = 'returns', cols_keep = NULL) {
+#' @importFrom xefun date_bop date_eop
+pq1_return = function(dt, x, freq='daily', num=1, date_type = 'eop', method = 'arithmetic', leading=TRUE, cumreturns = FALSE, rcol_name = NULL, cols_keep = c('symbol', 'date')) {
     . = Ra = chg = cumRa = byfreq = byyear = x_lag = NULL
     
     setkeyv(dt, "date")
     
     freq_list = c('daily', 'weekly', 'monthly', 'quarterly', 'yearly')
     if (freq == 'all') freq = freq_list
-    freq = intersect(freq_list, freq)
-    
-    freq = as.list(freq)
-    names(freq) = freq
+    freq = vec2lst(intersect(freq_list, freq))
     
     
         
@@ -50,36 +48,39 @@ pq1_return = function(dt, x, freq='monthly', num=1, date_type = 'eop', method = 
         ][, x_lag := shift(get(x), n=num, type = 'lag') 
         ][]
         if (leading) dateop = dateop[1, x_lag := datbop[1,get(x)]][]
-        dateop = dateop[, Ra := do.call(
+        
+        col_Ra = paste0('returns.',freqi)
+        dateop = dateop[, (col_Ra) := do.call(
             sprintf('return_%s', method), 
             args = list(x=get(x), x_lag)
-        )][, unique(c(cols_keep, 'Ra')), with = FALSE
-         ][, freq := freqi
+        )][, unique(c(cols_keep, col_Ra)), with = FALSE
          ]
         
+        col_cumRa = NULL 
         if (cumreturns == TRUE) {
+            col_cumRa = paste0('cumreturns.',freqi)
             dateop = dateop[
-                , chg := sum(Ra,1,na.rm = TRUE), keyby = c('symbol', 'date')
-            ][, cumRa := cumprod(chg), by = 'symbol'
+                , chg := sum(get(col_Ra),1,na.rm = TRUE), keyby = c('symbol', 'date')
+            ][, (col_cumRa) := cumprod(chg), by = 'symbol'
             ][, chg := NULL]
         }
         
         if (date_type == 'eop') {
-            dateop = dateop[, date := date_eop(date, freqi, workday = TRUE)]
+            dateop = dateop[, date := as_date(date_eop(freqi, date, workday = TRUE)) ][]
         } else if (date_type == 'bop') {
-            dateop = dateop[, date := date_bop(date, freqi, workday = TRUE)]
+            dateop = dateop[, date := as_date(date_bop(freqi, date, workday = TRUE)) ]
         }
         
         if (!is.null(rcol_name)) {
-            cols_rename = 'Ra'
-            if (cumreturns == TRUE) cols_rename = c('Ra', 'cumRa')
-            setnames(dateop, cols_rename, sub('Ra', rcol_name, cols_rename))
+            cols_from = c(col_Ra, col_cumRa)
+            cols_to = sub(col_Ra, rcol_name, cols_from)
+            setnames(dateop, cols_from, cols_to)
         }
         return(dateop)
     })
-
     
-    return(rbindlist(dt_rt)[])
+    dt_rt2 = Reduce(function(x,y)merge(x,y,all=TRUE), dt_rt)[]
+    return(dt_rt2)
 }
 
 
@@ -87,15 +88,15 @@ pq1_return = function(dt, x, freq='monthly', num=1, date_type = 'eop', method = 
 #' 
 #' \code{pq_return} calculates returns for daily series based on specified column, frequency and method type.
 #' 
-#' @param dt a list/dataframe of daily series dataset
+#' @param dt a list/dataframe of daily series.
 #' @param x the column name of adjusted asset price. 
-#' @param freq the frequency of returns. It supports c('all', 'daily', 'weekly', 'monthly', 'quarterly', 'yearly').
+#' @param freq the frequency of returns. It supports 'daily', 'weekly', 'monthly', 'quarterly', 'yearly' and 'all'. Defaults to daily.
 #' @param num the number of preceding periods used as the base value, defaults to 1, which means based on the previous period value.
 #' @param date_type the available date type are eop (end of period) and bop (beginning of period), defaults to the eop.
 #' @param method the method to calculate asset returns, the available methods including arithmetic and log, defaults to arithmetic. 
 #' @param leading whether to return the incomplete leading period returns.
 #' @param cumreturns logical, whether to return cumulative returns. Defaults to FALSE. 
-#' @param rcol_name setting the column name of returns, defaults to returns.
+#' @param rcol_name setting the column name of returns, defaults to NULL.
 #' @param cols_keep the columns keep in the return data. The columns of symbol, name and date will always kept if they are exist in the input data.
 #' @param date_range date range. Available value includes '1m'-'11m', 'ytd', 'max' and '1y'-'ny'. Default is max.
 #' @param from the start date. Default is NULL. If it is NULL, then calculate using date_range and end date.
@@ -118,27 +119,27 @@ pq1_return = function(dt, x, freq='monthly', num=1, date_type = 'eop', method = 
 #' 
 #' # cumulative returns
 #' dts_cumreturns = pq_return(datadj, x = 'close_adj', from = '2012-01-01', cumreturns = TRUE)
-#' e1 = pq_plot(dts_cumreturns, y = 'cumreturns', title='cumreturns', 
+#' e1 = pq_plot(dts_cumreturns, y = 'cumreturns.daily', title='cumreturns', 
 #'         arrange = list(rows=1, cols=1))
 #' e1[[1]]
 #' }
 #' 
 #' @export
-pq_return = function(dt, x, freq='monthly', num=1, date_type='eop', method='arithmetic', leading = TRUE, cumreturns = FALSE, rcol_name = 'returns', cols_keep=NULL, date_range='max', from=NULL, to=Sys.Date()) {
+pq_return = function(dt, x, freq='daily', num=1, date_type='eop', method='arithmetic', leading = TRUE, cumreturns = FALSE, rcol_name = NULL, cols_keep=NULL, date_range='max', from=NULL, to=Sys.Date()) {
     # arg
     ## method 
     method = check_arg(method, c('arithmetic', 'log'), default='arithmetic', arg_name = 'method')
     # arithmetic: p_{t-1} * (r+1) = p_{t}
     # log: p_{t-1} * e^r = p_{t}
     ## freq
-    freq = check_arg(freq, c('all', 'daily', 'weekly', 'monthly', 'quarterly', 'yearly'), default='all', arg_name = 'freq')
+    freq = check_arg(freq, c('daily', 'weekly', 'monthly', 'quarterly', 'yearly', 'all'), default='daily', arg_name = 'freq')
     # if (!check_freq_isdaily(dt)) stop('Please provide daily data')
     ## from/to
     to = check_to(to)
     from = check_from(date_range, from, to, default_from = "1000-01-01", default_date_range = 'max')
     
     ## dt
-    dt = check_dt(dt)
+    dt = check_dt(dt, symb_name = FALSE)
     ## price column 
     x = check_xcol(dt, x)
     ## kept columns